requirements.txt
abydos==0.5.0
altair==4.2.0
asttokens 
attrs==22.1.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.25.0
backcall 
blinker==1.5
cachetools==5.2.0
certifi 
charset-normalizer==2.1.0
click==8.1.3
colorama 
commonmark==0.9.1
cycler==0.11.0
debugpy 
decorator 
deprecation==2.1.0
entrypoints 
executing 
fonttools==4.36.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
ipykernel 
ipython 
ipywidgets==8.0.1
isodate==0.6.1
jedi 
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.12.1
jupyter-client 
jupyter-core 
jupyterlab-widgets==3.0.1
kiwisolver==1.4.4
language-tool-python==2.7.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline 
msrest==0.7.1
nest-asyncio 
nltk==3.7
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
parso 
pickleshare 
Pillow==9.2.0
prompt-toolkit 
protobuf==3.20.1
pure-eval 
pyarrow==9.0.0
pydeck==0.7.1
pyenchant==3.2.2
Pygments 
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil 
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pyzmq 
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
semver==2.13.0
six 
sklearn==0.0
smmap==5.0.0
stack-data 
streamlit==1.12.0
textblob==0.17.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado 
tqdm==4.64.0
traitlets 
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
watchdog==2.1.9
wcwidth 
widgetsnbextension==4.0.1
wincertstore==0.2
zipp==3.8.1
